GM Seed and Beta 7 changes
Update for GMSeed: there are two big problems since I originally posted
this, both related to optionals. One is that optionals are not
considered boolean values so

if myOptional{do someting} must be

if (myOptional != nil) {do something}</code>.

The second one is Apple is standardizing the API for what is and is not
optional. Much of what is optional has gone from implicit to explicit,
and some is no longer optional. IN particular UILabel and its property
text are optional, and navigationControllers property of
UIViewController is optional. What isn’t optional anymore is parts of
prepare for segue and much of the tableview parameters in the delegate.
 I've made the changes to the code here, but if you find any more, let
me know in a comment.